cv-assessment/requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
pycocotools>=2.0.7

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# OCR
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# Web Demo
gradio>=4.0.0

# Utilities
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
pyyaml>=6.0
